Fix Parquet UNKNOWN annotation IT writes on Dataproc
PyArrow writes to the local FS while spark_tmp_path is created on the
Hadoop default FS (GCS on Dataproc), so write the sample file locally
and copy_from_local into the test path.
